requirements.txt
openai-whisper
gradio
setuptools-rust
ffmpeg-python